Adding web-scraped doctor data
Education fields represent where the doctors went to school or had their residency. (Could use this and a ranked list of medical schools to find average school-score of a hospital).  
"Field" fields represent what the doctors specialize in (might be used if the user is looking for a particular type of doctor).  
"Hospital" fields represent the 1 or more hospitals that a doctor might work for