Added md score for each hospital
Based on education.  The first column is the hospital's foreign key, and the second column is the hospital's education score based on an average of each doctor that works for the hospital.

Uploading this now in case someone familiar with django wants to import it into the database.  If not, I'll figure it out tomorrow.